wire RateService into the API via the RateController

+++ src/main/scala/parkor/domain/Rates.scala
@@ -9,5 +9,5 @@ case class Rate(
   price: Price
 )
 
-case class Price(value: BigInt)
+case class Price(value: Int)
